requirements.txt
Flask
tensorflow
numpy
Pillow
werkzeug
